Creating ScheduleMeeting style file
